Format the Home Screen of T.O.P.
Realigned the html code to setup the homepage of The Odin Project to
match the desired outcome of the image.

Moved the header to the top and keep it there, similar to the footer.
Aligned the content of the homepage to the center of the screen regardless
of the size of the viewport.